Simplify xcm-config
ForeignCreatorsSovereignAccountOf is used by ForeignCreators filter to convert
location to AccountId, after ForeignCreators::IsForeign filter passes for an
(asset, location) pair.

The IsForeign filter is the actual differentiator, so if a location passes it,
we should support converting it to an AccountId.

As such, this commit replaces ForeignCreatorsSovereignAccountOf converter with
the more general LocationToAccountId converter.